requirements.txt.txt

pillow
pytesseract
opencv-python
tkinter

